reorganizing excitation stuff
